requirements.txt
joblib==1.1.0
pandas==1.3.5
pip==21.3.1
pyOpenSSL==21.0.0
pytest==7.1.2
requests-html==0.10.0
selenium==4.1.0
wheel==0.37.1
